requirements.txt
langchain 
langchain-community 
langchain-groq 
langchain-experimental 
neo4j 
wikipedia 
tiktoken 
yfiles_jupyter_graphs 
sentence-transformers
python-dotenv
streamlit
